+++ src/bda/plone/orders/browser/views.py
@@ -369,6 +369,16 @@ def customers_form_vocab():
     return [('', _('all', default='All'))] + customers
 
 
+def states_form_vocab():
+    states = vocabs.state_vocab()
+    return [('', _('all', default='All'))] + states.items()
+
+
+def salaried_form_vocab():
+    salaried = vocabs.salaried_vocab()
+    return [('', _('all', default='All'))] + salaried.items()
+
+
 class OrdersTable(OrdersTableBase):
 
     def render_filter(self):
@@ -402,15 +412,42 @@ class OrdersTable(OrdersTableBase):
                                default=u'Filter for customers'),
                 }
             )
-        if not vendor_selector and not customer_selector:
-            return
+
+        states = states_form_vocab()
+        state_selector = factory(
+            'label:select',
+            name='state',
+            value=self.request.form.get('state', ''),
+            props={
+                'vocabulary': states,
+                'label': _('filter_for_state',
+                           default=u'Filter for states'),
+            }
+        )
+
+        salaried = salaried_form_vocab()
+        salaried_selector = factory(
+            'label:select',
+            name='salaried',
+            value=self.request.form.get('salaried', ''),
+            props={
+                'vocabulary': salaried,
+                'label': _('filter_for_salaried',
+                           default=u'Filter for salaried state'),
+            }
+        )
+
         # concatenate filters
-        filter = ''
+        filter_widgets = ''
         if vendor_selector:
-            filter += vendor_selector(request=self.request)
+            filter_widgets += vendor_selector(request=self.request)
         if customer_selector:
-            filter += customer_selector(request=self.request)
-        return filter
+            filter_widgets += customer_selector(request=self.request)
+
+        filter_widgets += state_selector(request=self.request)
+        filter_widgets += salaried_selector(request=self.request)
+
+        return filter_widgets
 
     def render_order_actions_head(self):
         tag = Tag(Translate(self.request))
@@ -495,7 +532,9 @@ class OrdersTable(OrdersTableBase):
     def ajaxurl(self):
         params = [
             ('vendor', self.request.form.get('vendor')),
-            ('customer', self.request.form.get('customer'))
+            ('customer', self.request.form.get('customer')),
+            ('state', self.request.form.get('state')),
+            ('salaried', self.request.form.get('salaried')),
         ]
         query = urllib.urlencode(dict([it for it in params if it[1]]))
         query = query and '?{0}'.format(query) or ''
@@ -564,10 +603,22 @@ class OrdersData(OrdersTable, TableData):
             query = Any('vendor_uids', [vendor_uid])
         else:
             query = Any('vendor_uids', vendor_uids)
+
         # filter by customer if given
         customer = self.request.form.get('customer')
         if customer:
             query = query & Eq('creator', customer)
+
+        # Filter by state if given
+        state = self.request.form.get('state')
+        if state:
+            query = query & Eq('state', state)
+
+        # Filter by salaried if given
+        salaried = self.request.form.get('salaried')
+        if salaried:
+            query = query & Eq('salaried', salaried)
+
         # filter by search term if given
         term = self.request.form['sSearch'].decode('utf-8')
         if term: